feat(track-b): B4 — retire on-chain nametags + kill PROXY (Nostr-binding identity, D5)
Nametags are now Nostr bindings only (name<->chainPubkey); there is no on-chain nametag token and no PROXY addressing. Receive is always SignaturePredicate(chainPubkey).

Nametag retirement: registerNametag publishes the Nostr identity binding as the sole registration act (first-seen-wins is the uniqueness guard); the on-chain mint is dropped here and in the load / post-switch token-restore paths. Remove PaymentsModule.mintNametag + Sphere.mintNametag (public, breaking) and the NametagMinter module + its test; re-route Sphere.isNametagAvailable to transport.resolveNametag.

PROXY kill (send-path + identity + transport): resolveRecipientAddress drops PROXY emission — a nametag without a published DirectAddress now throws INVALID_RECIPIENT. Remove Sphere.getProxyAddress (public, breaking) + the proxy-address cache machinery. Remove PeerInfo.proxyAddress and stop publishing/deriving proxy addresses in the transport; AddressMode drops 'proxy'. Update affected tests (registerNametag, PaymentsModule nametag accessors, transport nametag resolution, 4 integration suites: drop the obsolete mint spy).

Deferred to B1 (entangled with the receiver gut): receive-path PROXY finalization (handleIncomingTransfer / InstantSplit*) and PROXY-parsing in core/address.ts + core/transport-resolver.ts.

Full suite green (2798 tests), tsc + lint clean.
