+++ modules/payments/index.ts
@@ -2,7 +2,6 @@ export * from './PaymentsModule';
 export * from './L1PaymentsModule';
 export * from './TokenSplitCalculator';
 export * from './TokenSplitExecutor';
-export * from './NametagMinter';
 export * from './TokenReservationLedger';
 export { SpendPlanner, SpendQueue, type ParsedTokenEntry, type ParsedTokenPool, type PlanResult } from './SpendQueue';
