+++ tests/integration/nametag-normalization.test.ts
@@ -106,7 +106,6 @@ function cleanTestDir(): void {
 describe('Nametag normalization integration', () => {
   let storage: FileStorageProvider;
   let tokenStorage: FileTokenStorageProvider;
-  let mintSpy: ReturnType<typeof vi.spyOn>;
 
   beforeEach(() => {
     cleanTestDir();
@@ -116,13 +115,9 @@ describe('Nametag normalization integration', () => {
     }
     storage = new FileStorageProvider({ dataDir: DATA_DIR });
     tokenStorage = new FileTokenStorageProvider({ tokensDir: TOKENS_DIR });
-    // Mock minting so registerNametag (mint-before-publish) succeeds without a real aggregator
-    mintSpy = vi.spyOn(Sphere.prototype as unknown as { mintNametag: () => Promise<unknown> }, 'mintNametag')
-      .mockResolvedValue({ success: true, token: null, nametagData: null });
   });
 
   afterEach(() => {
-    mintSpy.mockRestore();
     (Sphere as unknown as { instance: null }).instance = null;
     cleanTestDir();
     nostrRelayNametags.clear();
